# Add performance benchmark for shaping, get extents and draw