--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -226,7 +226,9 @@ jobs:
       # test a meson based dist
       - run: meson dist -Cbuild && rm -rf build
       # test experimental APIs
-      - run: meson build -Dexperimental_api=true && ninja -Cbuild test && rm -rf build
+      - run: meson build -Dexperimental_api=true -Dbenchmark=true -Dexperimental_api=true -Doptimization=2 && ninja -Cbuild test # or meson test -Cbuild
+      # run benchmarks
+      - run: build/perf/perf && meson test -Cbuild --benchmark && rm -rf build # or ninja -Cbuild benchmark
       # mingw
       - run: .ci/build-win32.sh
       - store_artifacts: