--- a/perf/perf.cc
+++ b/perf/perf.cc
@@ -1,0 +1,15 @@
+#include "benchmark/benchmark.h"
+
+#ifdef HAVE_CONFIG_H
+#include "config.h"
+#endif
+
+#include "perf-shaping.hh"
+#ifdef HAVE_FREETYPE
+#include "perf-extents.hh"
+#ifdef HB_EXPERIMENTAL_API
+#include "perf-draw.hh"
+#endif
+#endif
+
+BENCHMARK_MAIN ();