--- a/subprojects/.gitignore
+++ b/subprojects/.gitignore
@@ -6,3 +6,4 @@
 /proxy-libintl
 /zlib
 /packagecache
+/google-benchmark