`rm -rf build && meson build -Dbenchmark=true -Dexperimental_api=true -Doptimization=2 && ninja -Cbuild && build/perf/perf`

gives
```
---------------------------------------------------------------------------------------
Benchmark                                                Time           CPU Iterations
---------------------------------------------------------------------------------------
shape/fa-thelittleprince.txt - Amiri              72493871 ns   72339936 ns          9
shape/fa-thelittleprince.txt - NotoNastaliqUrdu  276136994 ns  275646177 ns          2
shape/fa-monologue.txt - Amiri                    12534512 ns   12508287 ns         58
shape/fa-monologue.txt - NotoNastaliqUrdu         49263441 ns   49196080 ns         13
shape/en-thelittleprince.txt - Roboto             34517598 ns   34475158 ns         20
shape/en-words.txt - Roboto                       40218073 ns   40162358 ns         17
extents/cff - ot - SourceSansPro                   2943239 ns    2939081 ns        238
extents/cff - ft - SourceSansPro                   7115167 ns    7105228 ns         97
extents/cff2 - ot - AdobeVFPrototype               2898143 ns    2894036 ns        241
extents/cff2 - ft - AdobeVFPrototype               2188817 ns    2186489 ns        317
extents/cff2/vf - ot - AdobeVFPrototype            3016606 ns    3013072 ns        232
extents/cff2/vf - ft - AdobeVFPrototype            2258966 ns    2256708 ns        309
extents/glyf - ot - SourceSerifVariable              39210 ns      39171 ns      18584
extents/glyf - ft - SourceSerifVariable            1062746 ns    1061467 ns        655
extents/glyf/vf - ot - SourceSerifVariable         4181210 ns    4176616 ns        168
extents/glyf/vf - ft - SourceSerifVariable         3770242 ns    3758127 ns        186
extents/glyf - ot - Comfortaa                        22834 ns      22734 ns      30695
extents/glyf - ft - Comfortaa                       614056 ns     611319 ns       1122
extents/glyf/vf - ot - Comfortaa                   2751386 ns    2738147 ns        259
extents/glyf/vf - ft - Comfortaa                   2348252 ns    2336775 ns        299
extents/glyf - ot - Roboto                           33721 ns      33570 ns      20162
extents/glyf - ft - Roboto                          764178 ns     760421 ns        912
draw/cff - ot - SourceSansPro                      3763109 ns    3744137 ns        188
draw/cff - ft - SourceSansPro                      7156332 ns    7120231 ns         96
draw/cff2 - ot - AdobeVFPrototype                  3042498 ns    3027313 ns        230
draw/cff2 - ft - AdobeVFPrototype                  2223677 ns    2213039 ns        316
draw/cff2/vf - ot - AdobeVFPrototype               3251870 ns    3236131 ns        216
draw/cff2/vf - ft - AdobeVFPrototype               2241302 ns    2230490 ns        314
draw/glyf - ot - SourceSerifVariable               3091201 ns    3076134 ns        228
draw/glyf - ft - SourceSerifVariable               1064737 ns    1059445 ns        655
draw/glyf/vf - ot - SourceSerifVariable            5414722 ns    5387867 ns        127
draw/glyf/vf - ft - SourceSerifVariable            3809064 ns    3788532 ns        174
draw/glyf - ot - Comfortaa                         1991267 ns    1981326 ns        350
draw/glyf - ft - Comfortaa                          618730 ns     615736 ns       1113
draw/glyf/vf - ot - Comfortaa                      3496356 ns    3479246 ns        201
draw/glyf/vf - ft - Comfortaa                      2381006 ns    2367965 ns        298
draw/glyf - ot - Roboto                            1777238 ns    1770355 ns        390
draw/glyf - ft - Roboto                             748675 ns     747645 ns        917
```

(comparing with freetype shouldn't be fair as the different API, linking status and optimization and things but is given as an illustration of difference between different functionalities. hb's `draw`, as provide close path and does some sanitization can be optimized easily)