@khaledhosny 2020-06-09
Hmm, I wonder why is Amiri ~2 times slower than NotoNastaliqUrdu.