@ebraminio 2020-06-09
> Hmm, I wonder why is Amiri ~2 times slower than NotoNastaliqUrdu.

Feel free to let me know any other test or combination could be tested here, thanks.