@ebraminio 2020-06-10
> Hmm, I wonder why is Amiri ~2 times slower than NotoNastaliqUrdu.

I was using a wrong blob as input 🤦 fixed now,