@khaledhosny 2020-06-11
> > Hmm, I wonder why is Amiri ~2 times slower than NotoNastaliqUrdu.
> 
> I was using a wrong blob as input 🤦 fixed now.

I see, you had me little worried!