@ebraminio 2020-06-24
> So, the first feedback: I want a commandline tool that takes parameters for what to test. It should take font file, text or glyph range or other info, as well as repeat count, fontfunc to use, etc. This should be a commandline tool.
> 
> The the test suit should specify the tests in a text file, like we do with our shaping test suite.
> 
> That's also similar to what `ftbench` does. Using Google perftool or something else, there was a package with a name like `grieger` but I cannot find (@garretrieger do you remember by any chance?!) that had code for the timing and statistics processing of reporting the timing. We need that in the commandline tool.

Is possible, but maybe we can have this first?