@behdad 2020-06-24
> Is possible, but maybe we can have this first?

Is you commit to finishing it to a reasonable state instead of just dropping into codebase and move on, yes.