@ebraminio 2020-06-24
> Is you commit to finishing it to a reasonable state instead of just dropping into codebase and move on, yes.

I did it the last time I added perf and now here we are, so :)