@ebraminio 2020-06-24
Just to add, have you seen
https://github.com/rhysd/github-action-benchmark
I actually have integrated this but disabled it as it write to gh-pages branch

Our https://github.com/harfbuzz/harfbuzz/blob/master/perf/run.sh even is able to achieve something similar, I will provide the tool and waiting for @garretrieger of course sometime as it is easy to get but let's don't encourage adhoc benchmarking and increase size of our benchmarking suite over the time instead encouraging adhoc ones, just like the project is doing with tests.

But please have a look at the first link first, it provides a similar experience to chromeperf thing we were dreaming about https://chromeperf.appspot.com/group_report?keys=agxzfmNocm9tZXBlcmZyFAsSB0Fub21hbHkYgIDIqI_uwQkM